Consolidated repeated code
To a reusable ‘formatTimeStamp’ function.
