+++ packages/lib-classifier/src/helpers/formatTimeStamp/formatTimeStamp.js
@@ -1,0 +1,17 @@
+function formatTimeStamp(seconds, duration) {
+  const currentVideoTime = duration ? seconds * duration : seconds
+
+  const pad = (string, digits) =>
+    ('0'.repeat(digits - 1) + string).slice(-digits)
+
+  const date = new Date(currentVideoTime * 1000)
+  const mm = pad(date.getUTCMinutes(), 2)
+  const ss = pad(date.getUTCSeconds(), 2)
+  const ms = pad(date.getUTCMilliseconds(), 3)
+  if (mm > 0) {
+    return `${mm}:${ss}:${ms}`
+  }
+  return `${ss}:${ms}`
+}
+
+export default formatTimeStamp
