+++ packages/lib-classifier/src/helpers/formatTimeStamp/index.js
@@ -1,0 +1,1 @@
+export { default } from './formatTimeStamp'
