+++ packages/lib-classifier/src/helpers/index.js
@@ -2,6 +2,7 @@ import * as featureDetection from './featureDetection'
 export { default as createLocationCounts } from './createLocationCounts'
 export { featureDetection }
 export { default as findLocationsByMediaType } from './findLocationsByMediaType'
+export { default as formatTimeStamp } from './formatTimeStamp'
 export { default as layouts } from './layouts'
 export { default as shownMarks } from './shownMarks'
 export { default as subjectViewers } from './subjectViewers'
