+++ packages/lib-classifier/src/plugins/drawingTools/models/marks/TemporalPoint/TemporalPoint.js
@@ -1,6 +1,7 @@
 import { getParentOfType, types } from 'mobx-state-tree'
 import { Point as PointComponent } from '@plugins/drawingTools/components'
 import { TemporalPointTool } from '@plugins/drawingTools/models/tools'
+import formatTimeStamp from '@helpers/formatTimeStamp'
 
 import Point from '../Point'
 
@@ -25,20 +26,7 @@ const TemporalPointModel = types
   .actions((self) => {
     function setVideoTime(displayTime, duration) {
       self.displayTime = displayTime
-
-      // creates 'ss:ms' timestamp
-      const currentVideoTime = displayTime * duration
-      const pad = (string, digits) =>
-        ('0'.repeat(digits - 1) + string).slice(-digits)
-
-      const date = new Date(currentVideoTime * 1000)
-      const mm = pad(date.getUTCMinutes(), 2)
-      const ss = pad(date.getUTCSeconds(), 2)
-      const ms = pad(date.getUTCMilliseconds(), 3)
-      if (mm > 0) {
-        self.displayTimeStamp = `${mm}:${ss}:${ms}`
-      }
-      self.displayTimeStamp = `${ss}:${ms}`
+      self.displayTimeStamp = formatTimeStamp(displayTime, duration)
     }
 
     return {
